requirements.txt
beautifulsoup4==4.12.3
pydantic==2.6.4
streamlit==1.32.2
openai==1.14.2
anthropic==0.21.3
google-generativeai==0.4.1
python-dotenv==1.0.1
aiohttp==3.9.3
requests==2.31.0
PyPDF2==3.0.1
